ENGR00234203 - Fix EGL initialization failure in Quadrant
Commit on behalf of Chingling Wang:
The reason for failing when eglInitialize() the second time
it was called is that the display is not reset back to
ELG_NO_DISPLAY in terminate().  It will not get egl
display(eglGetDisplay()) anymore, but the egl display is already
released in terminate(). So, invalid egl display will make
eglInitialize() fail.

Signed-off-by: Oliver Brown <oliver.brown@freescale.com>