MA-15744[Android_iMX8MQ_EVK]Touch:After setting dual display, some screens cannot be touched
When use two different resolution displays and show the same frame, the
frame size is larger than the smallest resolution display size. The
touch point position is normalized by mXScale and mYScale when are
calculated with the size of Surface. So the calculated touch point
position may be outside of the physical size when actual touch is in the
screen range.
Use the surface size to check if the pointer is isPointInsideSurface.

Change-Id: I5ebd9e245d0a1b6cfed837bd59e59730f4b892a9
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/10424
Reviewed-by: dandan.sun <dandan.sun@nxp.com>